+++ ibm_quantum_schemas/noise_learner_v3/version_0_2/__init__.py
@@ -1,0 +1,50 @@
+# This code is a Qiskit project.
+#
+# (C) Copyright IBM 2026.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""
+=============================================================================
+NoiseLearnerV3 v0.2 (:mod:`ibm_quantum_schemas.noise_learner_v3.version_0_2`)
+=============================================================================
+
+.. currentmodule:: ibm_quantum_schemas.noise_learner_v3.version_0_2
+
+Models for ``NoiseLearnerV3`` ``v0.2``.
+
+Classes
+=======
+
+.. autosummary::
+   :toctree: ../stubs/
+   :nosignatures:
+
+   ParamsModel
+   OptionsModel
+   PostSelectionOptionsModel
+   NoiseLearnerV3ResultsModel
+   NoiseLearnerV3ResultModel
+   TREXResultPostSelectionMetadataModel
+   TREXResultMetadataModel
+   LinbdbladResultPostSelectionMetadataModel
+   LinbdbladResultMetadataModel
+"""
+
+from ibm_quantum_schemas.noise_learner_v3.version_0_2.models import (
+    LinbdbladResultMetadataModel,
+    LinbdbladResultPostSelectionMetadataModel,
+    NoiseLearnerV3ResultModel,
+    NoiseLearnerV3ResultsModel,
+    OptionsModel,
+    ParamsModel,
+    PostSelectionOptionsModel,
+    TREXResultMetadataModel,
+    TREXResultPostSelectionMetadataModel,
+)